+++ src/messages.hpp
@@ -36,5 +36,11 @@
   "Please checkout " LIEF_DOC_PREFIX "/latest/extended/intro.html for the details"
 #endif
 
+#if !defined(NEEDS_EXTENDED_MSG)
+#define NEEDS_EXTENDED_MSG                                         \
+  "This function requires the extended version of LIEF.\n"                       \
+  "Please checkout " LIEF_DOC_PREFIX "/latest/extended/intro.html for the details"
+#endif
+
 
 #endif